Answer the right button, the trackpad and the keyboard the way each asks
Eight of the eleven complaints from a session with the app. Three of them were
features that already existed and were broken rather than missing.

A ring says where the next keystroke will go, which is a question you only have
while your hands are on the keyboard - so every primitive now reports what moved
the focus onto it and focus::ringVisible decides. Clicking a knob no longer
leaves it ringed in the colour the app also uses for selection.

mod-shift-scroll already expanded lanes and pitch rows, and did nothing at all on
a trackpad: a height is a whole pixel and a zoom is a factor, so lround(34 x
2^(0.002 x 3)) is 34 and every fragment rounded back to where it started.
RowView carries the fraction between events now. Two tests, both confirmed
failing against the old arithmetic.

The right button refused only the PRESS. juce::Button re-arms itself in
mouseDrag and completes the click in mouseUp, so a right-press that moved a
pixel fired every button in the application - the piano roll's octave pair is
where it was noticed. PopupPress states all three phases once, a gate refuses a
control that arms it without disarming it, and DewSlider does the same for the
one JUCE control dew builds on directly: a right-drag moved a knob, a fader and
both octave steppers, and wrote an undo step for it. The + and - an
IncDecButtons slider makes are children it creates itself, so they come through
createSliderButton.

The mixer opened two context menus. forwardChildMouseEventsTo is for hover, and
JUCE delivers a forwarded press to the listener as well as to the control, so a
right-click on a fader opened the parameter menu and the strip menu together.
The channel rack did the same through HeaderRow. isOwnPress is the other half of
the forwarder.

The Demos menu drew descriptions. menuRow packs a second line in with a newline
for drawPopupMenuItem to split, and the menu bar on macOS is the NATIVE one -
NSMenuItems no look and feel of dew's ever paints. Names only, a gate on any
translation unit that builds the menu bar, and the eight sentences nothing reads
any more are gone from the catalogue.

A remembered pattern the project has not got left the selector blank, the length
field disabled, the roll empty and the sequencer with no material. The rule
existed and ran on every path that replaced the document and on none that only
restored an id.

Volume and gain are the controls a person reaches for most, so they take the
app's own colour; funcLevel keeps saying "level" where a level is a signal
rather than a control, on the mixer's meter. Ruler numbers and the mixer's
routing labels move a rung up off the floor of the type scale. And the add-insert
button is a button at the head of a column rather than a bare + floating in one.

+++ src/ui/PianoRollView.cpp
@@ -263,7 +263,7 @@ void PianoRollComponent::zoomToFit()
     repaint();
 }
 
-void PianoRollComponent::setRowHeight (int wanted)
+void PianoRollComponent::setRowHeight (double wanted)
 {
     // The anchor, the clamp and the "it already fits" case are RowView's, and
     // the playlist's lanes run the same arithmetic. Ninety-seven rows at the
